Add AppConstants Issue #23
