+++ lib/pages/atoz.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'dart:async';
 import 'package:flutter_tts/flutter_tts.dart';
+import 'package:learn/utils/constants.dart';
 
 class ItemTile extends StatelessWidget {
   final int index;
@@ -22,19 +23,14 @@ class ItemTile extends StatelessWidget {
       color: item.backgroundColor,
       child: InkWell(
         onTap: () {
-          showDialog(
+          showModalBottomSheet(
             context: context,
-            // isScrollControlled: true,
+            isScrollControlled: true,
             builder: (BuildContext context) {
-              return Dialog(
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(25),
-                ),
-                child: _PopupContent(
-                  items: items,
-                  currentIndex: index,
-                  isAutoNextEnabled: isTimerEnabled,
-                ),
+              return _PopupContent(
+                items: items,
+                currentIndex: index,
+                isAutoNextEnabled: isTimerEnabled,
               );
             },
           );
@@ -54,13 +50,11 @@ class ItemTile extends StatelessWidget {
                 textAlign: TextAlign.center,
               ),
               const SizedBox(height: 3),
-              Expanded(
-                child: SvgPicture.asset(
-                  item.iconAsset,
-                  width: 100,
-                  height: 100,
-                  alignment: Alignment.center,
-                ),
+              SvgPicture.asset(
+                item.iconAsset,
+                width: 100,
+                height: 100,
+                alignment: Alignment.center,
               ),
               const SizedBox(height: 3),
               Text(item.description, textAlign: TextAlign.center),
@@ -147,96 +141,99 @@ class _PopupContentState extends State<_PopupContent> {
   @override
   Widget build(BuildContext context) {
     final currentItem = widget.items[currentIndex];
-    return Container(
-      width: MediaQuery.of(context).size.width * 0.7,
-      padding: const EdgeInsets.all(8),
-      decoration: BoxDecoration(
-        color: currentItem.backgroundColor,
-        borderRadius: BorderRadius.circular(25),
-      ),
-      child: SingleChildScrollView(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const SizedBox(height: 20),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(left: 5),
-                  child: Text(
-                    currentItem.title,
-                    style: const TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 40,
-                      color: Colors.black,
+    return SingleChildScrollView(
+      child: Container(
+        padding: EdgeInsets.only(
+          bottom: MediaQuery.of(context).viewInsets.bottom,
+        ),
+        child: Container(
+          color: Colors.transparent,
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 5),
+            child: SizedBox(
+              width: MediaQuery.of(context).size.width * 0.7,
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  const SizedBox(height: 20),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Padding(
+                        padding: const EdgeInsets.only(left: 5),
+                        child: Text(
+                          currentItem.title,
+                          style: const TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 40,
+                          ),
+                        ),
+                      ),
+                      IconButton(
+                        icon: const Icon(
+                          Icons.volume_up,
+                          size: 40,
+                        ),
+                        onPressed: () {
+                          _speakText(currentItem.title);
+                        },
+                      ),
+                    ],
+                  ),
+                  const SizedBox(height: 20),
+                  GestureDetector(
+                    onTap: () {
+                      _speakText(currentItem.description);
+                    },
+                    child: SvgPicture.asset(
+                      currentItem.iconAsset,
+                      width: 200,
+                      height: 200,
+                      alignment: Alignment.center,
                     ),
                   ),
-                ),
-                IconButton(
-                  icon: const Icon(
-                    Icons.volume_up,
-                    color: Colors.black,
-                    size: 40,
+                  const SizedBox(height: 20),
+                  Text(
+                    currentItem.description,
+                    textAlign: TextAlign.center,
+                    style: const TextStyle(
+                      fontSize: 28,
+                      // color: Colors.black,
+                    ),
                   ),
-                  onPressed: () {
-                    _speakText(currentItem.title);
-                  },
-                ),
-              ],
-            ),
-            const SizedBox(height: 20),
-            GestureDetector(
-              onTap: () {
-                _speakText(currentItem.description);
-              },
-              child: SvgPicture.asset(
-                currentItem.iconAsset,
-                width: 200,
-                height: 200,
-                alignment: Alignment.center,
+                  const SizedBox(height: 20),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                    children: [
+                      ElevatedButton(
+                        onPressed: _previousItem,
+                        child: const Text('Prev'),
+                      ),
+                      ElevatedButton(
+                        onPressed: _nextItem,
+                        child: const Text('Next'),
+                      ),
+                    ],
+                  ),
+                  const SizedBox(height: 20),
+                  ElevatedButton(
+                    style: ButtonStyle(
+                      backgroundColor: MaterialStateProperty.all(
+                          const Color.fromARGB(216, 233, 101, 92)),
+                    ),
+                    onPressed: () {
+                      Navigator.pop(context);
+                    },
+                    child: const Text(
+                      'Close',
+                      style: TextStyle(color: Colors.white),
+                    ),
+                  ),
+                  const SizedBox(height: 50),
+                ],
               ),
             ),
-            const SizedBox(height: 20),
-            Text(
-              currentItem.description,
-              textAlign: TextAlign.center,
-              style: const TextStyle(
-                fontSize: 28,
-                color: Colors.black,
-              ),
-            ),
-            const SizedBox(height: 20),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                ElevatedButton(
-                  onPressed: _previousItem,
-                  child: const Text('Prev'),
-                ),
-                ElevatedButton(
-                  onPressed: _nextItem,
-                  child: const Text('Next'),
-                ),
-              ],
-            ),
-            const SizedBox(height: 20),
-            ElevatedButton(
-              style: ButtonStyle(
-                backgroundColor: MaterialStateProperty.all(
-                    const Color.fromARGB(216, 233, 101, 92)),
-              ),
-              onPressed: () {
-                Navigator.pop(context);
-              },
-              child: const Text(
-                'Close',
-                style: TextStyle(color: Colors.white),
-              ),
-            ),
-            const SizedBox(height: 20),
-          ],
+          ),
         ),
       ),
     );
@@ -267,176 +264,18 @@ class AtoZ extends StatefulWidget {
 class _AtoZState extends State<AtoZ> {
   bool isTimerEnabled = false;
 
-  List<ItemData> items = [
-    ItemData(
-      iconAsset: 'assets/images/apple.svg',
-      title: 'A',
-      description: 'Apple',
-      backgroundColor: const Color.fromARGB(115, 171, 171, 171),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/ball.svg',
-      title: 'B',
-      description: 'Ball',
-      backgroundColor: const Color.fromARGB(115, 215, 118, 118),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/cat.svg',
-      title: 'C',
-      description: 'Cat',
-      backgroundColor: const Color.fromARGB(194, 130, 243, 69),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/dog.svg',
-      title: 'D',
-      description: 'Dog',
-      backgroundColor: const Color.fromARGB(115, 215, 199, 118),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/elephant.svg',
-      title: 'E',
-      description: 'Elephant',
-      backgroundColor: const Color.fromARGB(115, 118, 215, 173),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/fish.svg',
-      title: 'F',
-      description: 'Fish',
-      backgroundColor: const Color.fromARGB(115, 150, 118, 215),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/grapes.svg',
-      title: 'G',
-      backgroundColor: const Color.fromARGB(115, 215, 118, 175),
-      description: 'Grapes',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/horse.svg',
-      title: 'H',
-      backgroundColor: const Color.fromARGB(115, 157, 215, 118),
-      description: 'Horse',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/icecream.svg',
-      title: 'I',
-      backgroundColor: const Color.fromARGB(221, 176, 102, 220),
-      description: 'Ice-Cream',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/joker.svg',
-      title: 'J',
-      description: 'Joker',
-      backgroundColor: const Color.fromARGB(208, 112, 181, 206),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/king.svg',
-      title: 'K',
-      backgroundColor: const Color.fromARGB(115, 171, 215, 118),
-      description: 'King',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/lion.svg',
-      title: 'L',
-      description: 'Lion',
-      backgroundColor: const Color.fromARGB(236, 235, 229, 53),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/money.svg',
-      title: 'M',
-      description: 'Money',
-      backgroundColor: const Color.fromARGB(115, 118, 189, 215),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/nest.svg',
-      title: 'N',
-      description: 'Nest',
-      backgroundColor: const Color.fromARGB(115, 118, 215, 121),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/orange.svg',
-      title: 'O',
-      description: 'Orange',
-      backgroundColor: const Color.fromARGB(115, 215, 189, 118),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/parrot.svg',
-      title: 'P',
-      backgroundColor: const Color.fromARGB(115, 120, 118, 215),
-      description: 'Parrot',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/queen.svg',
-      title: 'Q',
-      backgroundColor: const Color.fromARGB(115, 215, 118, 118),
-      description: 'Queen',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/rabbit.svg',
-      title: 'R',
-      description: 'Rabbit',
-      backgroundColor: const Color.fromARGB(174, 134, 218, 191),
-    ),
-    ItemData(
-      iconAsset: 'assets/images/shiva.svg',
-      title: 'S',
-      backgroundColor: const Color.fromARGB(170, 156, 216, 145),
-      description: 'Shiva',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/table.svg',
-      title: 'T',
-      backgroundColor: const Color.fromARGB(180, 138, 64, 228),
-      description: 'Table',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/umbrella.svg',
-      title: 'U',
-      backgroundColor: const Color.fromARGB(189, 212, 127, 220),
-      description: 'Umbrella',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/van.svg',
-      title: 'V',
-      backgroundColor: const Color.fromARGB(115, 215, 118, 118),
-      description: 'Van',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/window.svg',
-      title: 'W',
-      backgroundColor: const Color.fromARGB(246, 255, 194, 25),
-      description: 'Window',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/xerox.svg',
-      title: 'X',
-      backgroundColor: const Color.fromARGB(115, 0, 236, 71),
-      description: 'Xerox',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/yellow.svg',
-      title: 'Y',
-      backgroundColor: const Color.fromARGB(115, 9, 255, 230),
-      description: 'Yellow',
-    ),
-    ItemData(
-      iconAsset: 'assets/images/zero.svg',
-      title: 'Z',
-      backgroundColor: const Color.fromARGB(155, 81, 0, 255),
-      description: 'Zero',
-    ),
-  ];
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
             const Text(
-              'A-Z',
+              AppConstants.a_z,
               style: TextStyle(fontWeight: FontWeight.bold),
             ),
-            const SizedBox(width: 10),
+            // const SizedBox(width: 10),
             Expanded(
               child: Align(
                 alignment: Alignment.centerRight,
@@ -466,10 +305,10 @@ class _AtoZState extends State<AtoZ> {
         child: GridView.count(
           crossAxisCount: 2,
           children: [
-            for (int i = 0; i < items.length; i++)
+            for (int i = 0; i < AppConstants.items.length; i++)
               ItemTile(
                 index: i,
-                items: items,
+                items: AppConstants.items,
                 isTimerEnabled: isTimerEnabled,
               ),
           ],
